default id gen

+++ txlcn-tm/src/main/java/com/codingapi/txlcn/tm/core/storage/FastStorage.java
@@ -200,5 +200,5 @@ public interface FastStorage {
      * @return
      * @throws FastStorageException
      */
-    int acquireOrRefreshMachineId(int machineId, int machineMaxSize, long timeout) throws FastStorageException;
+    int acquireOrRefreshMachineId(int machineId, long machineMaxSize, long timeout) throws FastStorageException;
 }
